use URI::Template to Level 2 RFC 6570 URI Template parameters.
Some API discovery documents (like Cloud Pub/Sub API) use level 2 url parameters like

"list": {
   "id": "pubsub.projects.topics.list",
   "path": "v1/{+project}/topics",
   "httpMethod": "GET",
...

The {+project} parameter above implies "Reserved Expansion". To support it and other
more complex URI parameters use the URI::Template per module.

Additionally, was getting an error when a body was not supplied in a request that
did not need a body, so check for that case.